Refactor: Member 도메인 CQS 패턴 적용
