+++ src/main/java/team/unibusk/backend/domain/member/application/command/MemberCommandService.java
@@ -1,24 +1,22 @@
-package team.unibusk.backend.domain.member.application;
+package team.unibusk.backend.domain.member.application.command;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import team.unibusk.backend.domain.member.application.dto.request.MemberNameUpdateServiceRequest;
-import team.unibusk.backend.domain.member.application.dto.response.MemberInfoResponse;
 import team.unibusk.backend.domain.member.application.dto.response.MemberNameUpdateResponse;
 import team.unibusk.backend.domain.member.domain.Member;
 import team.unibusk.backend.domain.member.domain.MemberRepository;
 
 @RequiredArgsConstructor
 @Service
-public class MemberService {
+public class MemberCommandService {
 
     private final MemberRepository memberRepository;
 
     @Transactional
     public MemberNameUpdateResponse updateMemberName(MemberNameUpdateServiceRequest request) {
         Member member = memberRepository.findByMemberId(request.memberId());
-
         member.updateName(request.name());
 
         return MemberNameUpdateResponse.builder()
@@ -26,13 +24,6 @@ public class MemberService {
                 .email(member.getEmail())
                 .name(member.getName())
                 .build();
-    }
-
-    @Transactional(readOnly = true)
-    public MemberInfoResponse getMyInfo(Long memberId) {
-        Member member = memberRepository.findByMemberId(memberId);
-
-        return MemberInfoResponse.from(member);
     }
 
 }
