+++ src/main/java/team/unibusk/backend/domain/member/presentation/command/MemberCommandController.java
@@ -1,35 +1,28 @@
-package team.unibusk.backend.domain.member.presentation;
+package team.unibusk.backend.domain.member.presentation.command;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import team.unibusk.backend.domain.member.application.MemberService;
-import team.unibusk.backend.domain.member.application.dto.response.MemberInfoResponse;
+import team.unibusk.backend.domain.member.application.command.MemberCommandService;
 import team.unibusk.backend.domain.member.application.dto.response.MemberNameUpdateResponse;
+import team.unibusk.backend.domain.member.presentation.MemberCommandDocsController;
 import team.unibusk.backend.domain.member.presentation.request.MemberNameUpdateRequest;
 import team.unibusk.backend.global.annotation.MemberId;
 
 @RequiredArgsConstructor
 @RequestMapping("/members")
 @RestController
-public class MemberController implements MemberDocsController {
+public class MemberCommandController implements MemberCommandDocsController {
 
-    private final MemberService memberService;
+    private final MemberCommandService memberCommandService;
 
     @PatchMapping("/me")
     public ResponseEntity<MemberNameUpdateResponse> updateMemberName(
             @MemberId Long memberId,
             @Valid @RequestBody MemberNameUpdateRequest request
     ) {
-        MemberNameUpdateResponse response = memberService.updateMemberName(request.toServiceRequest(memberId));
-
-        return ResponseEntity.status(200).body(response);
-    }
-
-    @GetMapping("/me")
-    public ResponseEntity<MemberInfoResponse> getMyInfo(@MemberId Long memberId) {
-        MemberInfoResponse response = memberService.getMyInfo(memberId);
+        MemberNameUpdateResponse response = memberCommandService.updateMemberName(request.toServiceRequest(memberId));
 
         return ResponseEntity.status(200).body(response);
     }
